MWPW-144110: rename merch-cards to merch-card-collection
update query index path resolution
fix unit tests
