+++ CODEOWNERS
@@ -39,7 +39,7 @@
 /libs/blocks/media/ @ryanmparrish
 /libs/blocks/merch/ @adobecom/tacocat
 /libs/blocks/merch-card/ @adobecom/tacocat
-/libs/blocks/merch-cards/ @adobecom/tacocat
+/libs/blocks/merch-card-collection/ @adobecom/tacocat
 /libs/blocks/merch-offers/ @adobecom/tacocat
 /libs/blocks/mnemonic-list/ @adobecom/tacocat
 /libs/blocks/ost/ @adobecom/tacocat
