+++ libs/blocks/merch-card-collection/merch-card-collection.css
@@ -1,0 +1,3 @@
+.section[class*="-merch-cards"] .full-width {
+  grid-column: 1 / -1;
+}
