--- libs/blocks/merch-cards/merch-cards.css
@@ -1,15 +1,0 @@
-.section[class*="-merch-cards"] .full-width {
-  grid-column: 1 / -1;
-}
-
-merch-cards {
-  min-height: 200px;
-}
-
-merch-cards.loading {
-  visibility: hidden;
-}
-
-merch-card.placeholder {
-  visibility: hidden;
-}
