--- libs/blocks/merch-cards/merch-cards.js
@@ -1,320 +1,0 @@
-import '../../deps/merch-cards.js';
-import '../../deps/merch-card.js';
-import '../../deps/commerce.js';
-import '../merch/merch.js';
-import '../merch-card/merch-card.js';
-import { createTag, decorateLinks, getConfig, loadBlock, loadStyle } from '../../utils/utils.js';
-import { replaceText } from '../../features/placeholders.js';
-import { OVERRIDE_PATHS } from '../../features/personalization/personalization.js';
-
-const DIGITS_ONLY = /^\d+$/;
-
-const CLASS_LOADING = 'loading';
-
-const LITERAL_SLOTS = [
-  'searchText',
-  'filtersText',
-  'sortText',
-  'popularityText',
-  'alphabeticallyText',
-  'noResultText',
-  'resultText',
-  'resultsText',
-  'searchResultText',
-  'searchResultsText',
-  'searchResultMobileText',
-  'searchResultsMobileText',
-  'noSearchResultsText',
-  'noSearchResultsMobileText',
-  'showMoreText',
-];
-
-// allows improve TBT by returning control to the main thread.
-// eslint-disable-next-line no-promise-executor-return
-const makePause = async () => new Promise((resolve) => setTimeout(resolve, 0));
-
-const fail = (el, err = '') => {
-  window.lana?.log(`Failed to initialize merch cards: ${err}`);
-  el.innerHTML = '';
-  return el;
-};
-
-/**
- * Removes merch cards from the DOM if they are not meant to be displayed for the given filter.
- * @param {*} merchCards merch-cards element
- */
-export function filterMerchCards(merchCards, filtered) {
-  [...merchCards.children].filter((child) => child.tagName === 'MERCH-CARD')
-    .forEach((card) => {
-      if (!Object.prototype.hasOwnProperty.call(card.filters, filtered)) {
-        card.remove();
-      }
-    });
-}
-
-/** parse card preferences for each category filter */
-export function parsePreferences(elements) {
-  return [...elements].map((el) => {
-    let size;
-    if (el?.firstElementChild?.tagName === 'STRONG') {
-      size = 'wide';
-      if (el.firstElementChild.firstElementChild?.tagName === 'U') {
-        size = 'super-wide';
-      }
-    }
-    return [el.textContent?.trim(), size];
-  });
-}
-
-const localizedPath = (path, config) => `${config?.locale?.prefix && path.indexOf(config.locale.prefix) !== 0 ? config.locale.prefix : ''}${path}`;
-
-const fetchOverrideCard = (path, config) => new Promise((resolve, reject) => {
-  try {
-    fetch(`${localizedPath(path, config)}.plain.html`).then((res) => {
-      if (res.ok) {
-        res.text().then((cardContent) => {
-          resolve({ path, cardContent: /^<div>(.*)<\/div>$/.exec(cardContent.replaceAll('\n', ''))[1] });
-        });
-      } else {
-        reject(res.statusText || res.status);
-      }
-    });
-  } catch (error) {
-    reject(error);
-  }
-});
-
-async function generateCards(html, config) {
-  const cards = `<div>${html}</div>`;
-  const fragment = document.createRange().createContextualFragment(cards);
-  const cardsRoot = fragment.firstElementChild;
-  await makePause();
-  // Replace placeholders
-  cardsRoot.innerHTML = await replaceText(cardsRoot.innerHTML, config);
-  await makePause();
-  const autoBlocks = await decorateLinks(cardsRoot).map(loadBlock);
-  await Promise.all(autoBlocks);
-  await makePause();
-
-  const blocks = [...cardsRoot.querySelectorAll(':scope > div')].map(loadBlock);
-
-  // process merch card blocks in batches of 3.
-  const batchSize = 3;
-  for (let i = 0; i < blocks.length; i += batchSize) {
-    const batch = blocks.slice(i, i + batchSize);
-    await Promise.all(batch);
-    await makePause();
-  }
-  return cardsRoot;
-}
-
-async function overrideCards(root, overridePromises, config) {
-  try {
-    if (overridePromises?.length > 0) {
-      // Wait for all override cards to be fetched
-      const overrideData = await Promise.all(overridePromises);
-
-      if (overrideData?.length > 0) {
-        const overrideRoot = await generateCards(overrideData?.map(({ cardContent }) => cardContent).join('\n'), config);
-        const overrideMap = {};
-        overrideRoot.querySelectorAll('merch-card').forEach((card) => { if (card.name) { overrideMap[card.name] = card; } });
-        root.querySelectorAll('merch-card').forEach((card) => {
-          if (card.name && overrideMap[card.name]) {
-            card.replaceWith(overrideMap[card.name]);
-          }
-        });
-      }
-    }
-  } catch (error) {
-    window?.lana.error('Failed to override cards', error);
-  }
-  return root;
-}
-
-async function initMerchCards(config, type, filtered, el, preferences, overrides) {
-  let cardsData;
-  let err;
-
-  const overridePromises = overrides?.split(',').map(fetchOverrideCard);
-
-  try {
-    const url = `${localizedPath(config.queryIndexCardPath, config)}.json?sheet=${type}`;
-    const res = await fetch(url);
-    if (res.ok) {
-      cardsData = await res.json();
-    } else {
-      err = res.statusText || res.status;
-    }
-  } catch (error) {
-    err = error.message;
-  }
-  if (!cardsData?.data) {
-    fail(el, err);
-  }
-
-  let cardsRoot = await generateCards(cardsData.data.map(({ cardContent }) => cardContent).join('\n'), config);
-
-  cardsRoot = await overrideCards(cardsRoot, overridePromises, config);
-
-  if (filtered) {
-    filterMerchCards(cardsRoot, filtered);
-  }
-  // re-order cards, update card filters
-  [...cardsRoot.children].filter((card) => card.tagName === 'MERCH-CARD').forEach((merchCard) => {
-    const filters = { ...merchCard.filters };
-    Object.keys(filters).forEach((key) => {
-      const preference = preferences[key];
-      if (!preference) return;
-      preference
-        .forEach(([sortKey, cardSize], index) => {
-          if (merchCard.name === sortKey || merchCard.title === sortKey) {
-            filters[key] = { order: index + 1, size: cardSize };
-          }
-        });
-    });
-    merchCard.filters = filters;
-  });
-  return cardsRoot;
-}
-
-export default async function init(el) {
-  if (el.classList.length < 2) {
-    return fail(el, 'Missing collection type');
-  }
-
-  const overrides = el.dataset[OVERRIDE_PATHS];
-  const config = getConfig();
-  if (!config.queryIndexCardPath) {
-    return fail(el, 'Missing queryIndexCardPath config');
-  }
-
-  const { base } = getConfig();
-  const merchStyles = new Promise((resolve) => {
-    loadStyle(`${base}/blocks/merch/merch.css`, resolve);
-  });
-  const merchCardStyles = new Promise((resolve) => {
-    loadStyle(`${base}/blocks/merch-card/merch-card.css`, resolve);
-  });
-  const allStyles = Promise.all([merchStyles, merchCardStyles]);
-
-  const attributes = { filter: 'all', class: CLASS_LOADING };
-  const settingsEl = el.firstElementChild?.firstElementChild;
-
-  const filtered = settingsEl?.firstElementChild?.tagName === 'STRONG';
-
-  if (!filtered) {
-    await Promise.all([
-      import(`${base}/features/spectrum-web-components/dist/theme.js`),
-      import(`${base}/features/spectrum-web-components/dist/button.js`),
-      import(`${base}/features/spectrum-web-components/dist/search.js`),
-      import(`${base}/features/spectrum-web-components/dist/overlay.js`),
-      import(`${base}/features/spectrum-web-components/dist/menu.js`),
-      import(`${base}/features/spectrum-web-components/dist/popover.js`),
-    ]);
-  }
-
-  const preferences = {};
-
-  if (settingsEl) {
-    const [filter, limit = 24] = settingsEl
-      .textContent.split(',').map((s) => s.trim());
-
-    settingsEl.parentElement.remove();
-
-    if (filtered) {
-      attributes.filtered = filter;
-      preferences[filter] = parsePreferences(el.querySelectorAll('p'));
-    } else {
-      if (filter) {
-        attributes.filter = filter;
-      }
-      [...el.children].forEach((filterElement) => {
-        const filterName = filterElement.firstElementChild.textContent?.trim();
-        const elements = [...filterElement.lastElementChild.querySelectorAll('p')];
-        preferences[filterName] = parsePreferences(elements);
-      });
-    }
-
-    if (limit) {
-      attributes.limit = limit;
-    }
-  }
-
-  const literalsEl = el.lastElementChild?.firstElementChild;
-  // parse literals
-  const literalSlots = [];
-  if (literalsEl && /filter/.test(literalsEl.querySelector('u')?.innerText)) {
-    literalsEl.querySelectorAll('u').forEach((u) => {
-      const text = u.innerText.trim();
-      if (DIGITS_ONLY.test(text)) {
-        u.outerHTML = '<span data-placeholder="resultCount"></span>';
-      } else if (text === 'search') {
-        u.outerHTML = '<span data-placeholder="searchTerm"></span>';
-      } else if (text === 'filter') {
-        u.outerHTML = '<span data-placeholder="filter"></span>';
-      }
-    });
-    let index = 0;
-    while (literalsEl.firstElementChild) {
-      const literalEl = literalsEl.firstElementChild;
-      let slot;
-      if (literalEl.tagName === 'P') {
-        slot = literalEl;
-      } else if (literalEl.tagName === 'HR') {
-        slot = createTag('div');
-        while (literalEl.nextElementSibling && literalEl.nextElementSibling?.tagName !== 'HR') {
-          slot.appendChild(literalEl.nextElementSibling);
-        }
-        literalEl.nextElementSibling?.remove?.();
-      }
-      literalEl.remove();
-      if (slot) {
-        slot.setAttribute('slot', LITERAL_SLOTS[index]);
-        index += 1;
-      }
-      literalSlots.push(slot);
-    }
-  }
-
-  const type = el.classList[1];
-  const merchCards = createTag('merch-cards', attributes);
-  if (literalSlots.length > 0) {
-    merchCards.append(...literalSlots);
-  } else if (!merchCards.filtered) {
-    merchCards.filtered = 'all';
-  }
-  await initMerchCards(config, type, attributes.filtered, el, preferences, overrides)
-    .then((async (cardsRoot) => {
-      const cards = [...cardsRoot.children];
-      const batchSize = 3;
-      for (let i = 0; i < cards.length; i += batchSize) {
-        const batch = cards.slice(i, i + batchSize);
-        merchCards.append(...batch);
-        if (i === batchSize) {
-          // avoid layout shift due to size of first cards.
-          merchCards.requestUpdate();
-        }
-        await makePause();
-      }
-      merchCards.updateComplete.then(() => {
-        merchCards.classList.remove(CLASS_LOADING);
-      });
-      merchCards.displayResult = true;
-    }
-    ));
-
-  const appContainer = document.querySelector('.merch.app');
-
-  if (appContainer) {
-    merchCards.classList.add('four-merch-cards', type);
-    appContainer.appendChild(merchCards);
-    el.remove();
-  } else {
-    if (!el.closest('main > .section[class*="-merch-card"]')) {
-      el.closest('main > .section').classList.add('four-merch-cards', type);
-    }
-    el.replaceWith(merchCards);
-  }
-  await allStyles;
-  return merchCards;
-}
