+++ test/blocks/merch-card-collection/mocks/merch-card-collection.html
@@ -10,23 +10,31 @@
     </div>
 
     <div id="defaultFilter" class="merch-cards catalog">
+      <div>
+        <a href="https://main--milo--adobecom.hlx.live/query-index-cards.json?sheet=catalog">query index</a>
+      </div>
     </div>
 
     <div id="filterShowMoreText" class="merch-cards catalog">
       <div>
-        <div>all</div>
+        <a href="https://main--milo--adobecom.hlx.live/query-index-cards.json?sheet=catalog">query index</a>
+        <p>all</p>
       </div>
     </div>
 
     <div id="filterShowMoreTextLimit" class="merch-cards catalog">
       <div>
-        <div>all, 9</div>
+        <a href="https://main--milo--adobecom.hlx.live/query-index-cards.json?sheet=catalog">query index</a>
+        <p>all, 9</p>
       </div>
     </div>
 
     <div id="onlyPhoto" class="merch-cards catalog">
       <div>
-        <div><strong>photo</strong></div>
+        <div>
+          <a href="https://main--milo--adobecom.hlx.live/query-index-cards.json?sheet=catalog">query index</a>
+          <p><strong>photo</strong></p>
+        </div>
       </div>
       <div>
         <div>
@@ -39,7 +47,10 @@
 
     <div id="multipleFilters" class="merch-cards catalog">
       <div>
-        <div>all, 9</div>
+        <div>
+          <a href="https://main--milo--adobecom.hlx.live/query-index-cards.json?sheet=catalog">query index</a>
+          <p>all, 9</p>
+        </div>
       </div>
       <div>
         <div>all</div>
@@ -94,7 +105,8 @@
 
     <div id="placeholders" class="merch-cards catalog">
       <div>
-        <div>all, 18</div>
+        <a href="https://main--milo--adobecom.hlx.live/query-index-cards.json?sheet=catalog">query index</a>
+        <p>all, 18</p>
       </div>
       <div>
         <div>
@@ -125,10 +137,10 @@
     </div>
 
     <!--this the expected merch-cards element -->
-    <merch-cards filter="all" class="" limit="18" types="" search="" page="1"><p slot="searchText">Search all products</p><p slot="filtersText">Filters</p><p slot="sortText">Sort</p><p slot="popularityText">Popularity</p><p slot="alphabeticallyText">Alphabetical</p><p slot="noResultText">0 results</p><p slot="resultText">1 result in <strong><span data-placeholder="filter"></span></strong></p><p slot="resultsText"><span data-placeholder="resultCount"></span> results in <strong><span data-placeholder="filter"></span></strong></p><p slot="searchResultText">1 result for <strong><span data-placeholder="searchTerm"></span></strong></p><p slot="searchResultsText"><span data-placeholder="resultCount"></span> results for <strong><span data-placeholder="searchTerm"></span></strong></p><p slot="searchResultMobileText">1 result for: <strong><span data-placeholder="searchTerm"></span></strong></p><p slot="searchResultsMobileText"><span data-placeholder="resultCount"></span> results for: <strong><span data-placeholder="searchTerm"></span></strong></p><p slot="noSearchResultsText">Your search for <strong><span data-placeholder="searchTerm"></span></strong> did not yield any results</p><div slot="noSearchResultsMobileText"><p>Your search for <strong><span data-placeholder="searchTerm"></span></strong> did not yield any results. Try a different search term.</p><p>Suggestions:</p><ul>
+    <merch-card-collection filter="all" class="merch-cards catalog" limit="18" types="" search="" page="1"><p slot="searchText">Search all products</p><p slot="filtersText">Filters</p><p slot="sortText">Sort</p><p slot="popularityText">Popularity</p><p slot="alphabeticallyText">Alphabetical</p><p slot="noResultText">0 results</p><p slot="resultText">1 result in <strong><span data-placeholder="filter"></span></strong></p><p slot="resultsText"><span data-placeholder="resultCount"></span> results in <strong><span data-placeholder="filter"></span></strong></p><p slot="searchResultText">1 result for <strong><span data-placeholder="searchTerm"></span></strong></p><p slot="searchResultsText"><span data-placeholder="resultCount"></span> results for <strong><span data-placeholder="searchTerm"></span></strong></p><p slot="searchResultMobileText">1 result for: <strong><span data-placeholder="searchTerm"></span></strong></p><p slot="searchResultsMobileText"><span data-placeholder="resultCount"></span> results for: <strong><span data-placeholder="searchTerm"></span></strong></p><p slot="noSearchResultsText">Your search for <strong><span data-placeholder="searchTerm"></span></strong> did not yield any results</p><div slot="noSearchResultsMobileText"><p>Your search for <strong><span data-placeholder="searchTerm"></span></strong> did not yield any results. Try a different search term.</p><p>Suggestions:</p><ul>
             <li>Make sure all words are spelled correctly</li>
             <li>Use quotes to search for an entire phrase, such as "crop an image"</li>
-          </ul></div><p slot="showMoreText">Show more</p></merch-cards>
+          </ul></div><p slot="showMoreText">Show more</p></merch-card-collection>
   </div>
 
   <div id="cards">
